+++ src/main/java/leets/weeth/global/common/controller/ExceptionDocController.java
@@ -1,4 +1,4 @@
-package leets.weeth.global.presentation;
+package leets.weeth.global.common.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
