+++ src/main/java/leets/weeth/global/common/exception/ErrorCodeInterface.java
@@ -10,6 +10,7 @@ public interface ErrorCodeInterface {
     HttpStatus getStatus();
     String getMessage();
 
+    // ExplainError 어노테이션에 작성된 설명을 조회하는 메서드
     default String getExplainError() throws NoSuchFieldException {
         Field field = this.getClass().getField(((Enum<?>) this).name());
         ExplainError annotation = field.getAnnotation(ExplainError.class);
